Remove unnecessary INotifyPropertyChanged interface from DisposeObject.
