+++ Source/HelixToolkit.SharpDX.Shared/Core/Abstract/RenderCore.cs
@@ -275,10 +275,9 @@ namespace HelixToolkit.UWP
             /// <typeparam name="T"></typeparam>
             /// <param name="backingField"></param>
             /// <param name="value"></param>
-            /// <param name="propertyName"></param>
             /// <returns></returns>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            protected bool SetAffectsRender<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+            protected bool SetAffectsRender<T>(ref T backingField, T value)
             {
                 if (EqualityComparer<T>.Default.Equals(backingField, value))
                 {
@@ -286,7 +285,6 @@ namespace HelixToolkit.UWP
                 }
 
                 backingField = value;
-                this.RaisePropertyChanged(propertyName);
                 RaiseInvalidateRender();
                 return true;
             }
@@ -297,10 +295,9 @@ namespace HelixToolkit.UWP
             /// <typeparam name="T"></typeparam>
             /// <param name="backingField">The backing field.</param>
             /// <param name="value">The value.</param>
-            /// <param name="propertyName">Name of the property.</param>
             /// <returns></returns>
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            protected bool SetAffectsCanRenderFlag<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+            protected bool SetAffectsCanRenderFlag<T>(ref T backingField, T value)
             {
                 if (EqualityComparer<T>.Default.Equals(backingField, value))
                 {
@@ -308,7 +305,6 @@ namespace HelixToolkit.UWP
                 }
 
                 backingField = value;
-                this.RaisePropertyChanged(propertyName);
                 UpdateCanRenderFlag();
                 RaiseInvalidateRender();
                 return true;
