+++ Source/HelixToolkit.SharpDX.Shared/Core2D/Abstract/RenderCore2D.cs
@@ -206,9 +206,8 @@ namespace HelixToolkit.UWP
             /// <typeparam name="T"></typeparam>
             /// <param name="backingField"></param>
             /// <param name="value"></param>
-            /// <param name="propertyName"></param>
             /// <returns></returns>
-            protected bool SetAffectsRender<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+            protected bool SetAffectsRender<T>(ref T backingField, T value)
             {
                 if (EqualityComparer<T>.Default.Equals(backingField, value))
                 {
@@ -216,7 +215,6 @@ namespace HelixToolkit.UWP
                 }
 
                 backingField = value;
-                this.RaisePropertyChanged(propertyName);
                 InvalidateRenderer();
                 return true;
             }
