+++ Source/HelixToolkit.SharpDX.Shared/Model/Scene/Abstract/SceneNode.cs
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Collections.ObjectModel;
-
+using System.ComponentModel;
 #if !NETFX_CORE
 namespace HelixToolkit.Wpf.SharpDX
 #else
@@ -23,12 +23,12 @@ namespace HelixToolkit.UWP
     namespace Model.Scene
     {
         using Core;
-        using Render;       
-
+        using Render;
+        
         /// <summary>
         ///
         /// </summary>
-        public abstract partial class SceneNode : DisposeObject, IComparable<SceneNode>
+        public abstract partial class SceneNode : DisposeObject, IComparable<SceneNode>, INotifyPropertyChanged
         {
             #region Properties
 
@@ -886,6 +886,102 @@ namespace HelixToolkit.UWP
                 base.OnDispose(disposeManagedResources);
             }
 
+            public int CompareTo(SceneNode other)
+            {
+                if(other == null) { return 1; }
+                return RenderOrderKey.CompareTo(other.RenderOrderKey);
+            }
+
+            public void RaiseMouseDownEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
+            {
+                MouseDown?.Invoke(this, new SceneNodeMouseDownArgs(viewport, pos, this, hit));
+            }
+
+            public void RaiseMouseMoveEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
+            {
+                MouseMove?.Invoke(this, new SceneNodeMouseMoveArgs(viewport, pos, this, hit));
+            }
+
+            public void RaiseMouseUpEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
+            {
+                MouseUp?.Invoke(this, new SceneNodeMouseUpArgs(viewport, pos, this, hit));
+            }
+
+            #region INotifyPropertyChanged
+            private bool disablePropertyChangedEvent = false;
+            /// <summary>
+            /// Disable property changed event calling
+            /// </summary>
+            public bool DisablePropertyChangedEvent
+            {
+                set
+                {
+                    if (disablePropertyChangedEvent == value)
+                    {
+                        return;
+                    }
+                    disablePropertyChangedEvent = value;
+                    RaisePropertyChanged();
+                }
+                get
+                {
+                    return disablePropertyChangedEvent;
+                }
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            public event PropertyChangedEventHandler PropertyChanged;
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <param name="propertyName"></param>
+            protected void RaisePropertyChanged([CallerMemberName] string propertyName = "")
+            {
+                if (!DisablePropertyChangedEvent)
+                    PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <typeparam name="T"></typeparam>
+            /// <param name="backingField"></param>
+            /// <param name="value"></param>
+            /// <param name="propertyName"></param>
+            /// <returns></returns>
+            protected bool Set<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+            {
+                if (EqualityComparer<T>.Default.Equals(backingField, value))
+                {
+                    return false;
+                }
+
+                backingField = value;
+                this.RaisePropertyChanged(propertyName);
+                return true;
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <typeparam name="T"></typeparam>
+            /// <param name="backingField"></param>
+            /// <param name="value"></param>
+            /// <param name="raisePropertyChanged"></param>
+            /// <param name="propertyName"></param>
+            /// <returns></returns>
+            protected bool Set<T>(ref T backingField, T value, bool raisePropertyChanged, [CallerMemberName] string propertyName = "")
+            {
+                if (EqualityComparer<T>.Default.Equals(backingField, value))
+                {
+                    return false;
+                }
+
+                backingField = value;
+                if (raisePropertyChanged)
+                { this.RaisePropertyChanged(propertyName); }
+                return true;
+            }
+
             /// <summary>
             /// 
             /// </summary>
@@ -927,27 +1023,7 @@ namespace HelixToolkit.UWP
                 InvalidateSceneGraph();
                 return true;
             }
-
-            public int CompareTo(SceneNode other)
-            {
-                if(other == null) { return 1; }
-                return RenderOrderKey.CompareTo(other.RenderOrderKey);
-            }
-
-            public void RaiseMouseDownEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
-            {
-                MouseDown?.Invoke(this, new SceneNodeMouseDownArgs(viewport, pos, this, hit));
-            }
-
-            public void RaiseMouseMoveEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
-            {
-                MouseMove?.Invoke(this, new SceneNodeMouseMoveArgs(viewport, pos, this, hit));
-            }
-
-            public void RaiseMouseUpEvent(IViewport3DX viewport, Vector2 pos, HitTestResult hit)
-            {
-                MouseUp?.Invoke(this, new SceneNodeMouseUpArgs(viewport, pos, this, hit));
-            }
+            #endregion
         }
 
         public sealed class NullSceneNode : SceneNode
