+++ Source/HelixToolkit.SharpDX.Shared/Model/Scene2D/Abstract/SceneNode2D.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Collections.ObjectModel;
+using System.ComponentModel;
 #if !NETFX_CORE
 namespace HelixToolkit.Wpf.SharpDX
 #else
@@ -27,7 +28,7 @@ namespace HelixToolkit.UWP
         /// <summary>
         ///
         /// </summary>
-        public abstract partial class SceneNode2D : DisposeObject, IHitable2D
+        public abstract partial class SceneNode2D : DisposeObject, IHitable2D, INotifyPropertyChanged
         {
             public sealed class UpdateEventArgs : EventArgs
             {
@@ -566,6 +567,83 @@ namespace HelixToolkit.UWP
                 base.OnDispose(disposeManagedResources);
             }
 
+
+
+            #region INotifyPropertyChanged
+            private bool disablePropertyChangedEvent = false;
+            /// <summary>
+            /// Disable property changed event calling
+            /// </summary>
+            public bool DisablePropertyChangedEvent
+            {
+                set
+                {
+                    if (disablePropertyChangedEvent == value)
+                    {
+                        return;
+                    }
+                    disablePropertyChangedEvent = value;
+                    RaisePropertyChanged();
+                }
+                get
+                {
+                    return disablePropertyChangedEvent;
+                }
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            public event PropertyChangedEventHandler PropertyChanged;
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <param name="propertyName"></param>
+            protected void RaisePropertyChanged([CallerMemberName] string propertyName = "")
+            {
+                if (!DisablePropertyChangedEvent)
+                    PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <typeparam name="T"></typeparam>
+            /// <param name="backingField"></param>
+            /// <param name="value"></param>
+            /// <param name="propertyName"></param>
+            /// <returns></returns>
+            protected bool Set<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+            {
+                if (EqualityComparer<T>.Default.Equals(backingField, value))
+                {
+                    return false;
+                }
+
+                backingField = value;
+                this.RaisePropertyChanged(propertyName);
+                return true;
+            }
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <typeparam name="T"></typeparam>
+            /// <param name="backingField"></param>
+            /// <param name="value"></param>
+            /// <param name="raisePropertyChanged"></param>
+            /// <param name="propertyName"></param>
+            /// <returns></returns>
+            protected bool Set<T>(ref T backingField, T value, bool raisePropertyChanged, [CallerMemberName] string propertyName = "")
+            {
+                if (EqualityComparer<T>.Default.Equals(backingField, value))
+                {
+                    return false;
+                }
+
+                backingField = value;
+                if (raisePropertyChanged)
+                { this.RaisePropertyChanged(propertyName); }
+                return true;
+            }
+
             /// <summary>
             ///
             /// </summary>
@@ -607,6 +685,7 @@ namespace HelixToolkit.UWP
                 InvalidateMeasure();
                 return true;
             }
+            #endregion
         }
     }
